Create Silent OneDrive Backup (RMM Deployment)
This script will run from the System and impersonate the logged in user using an RMM for deployment.

This only works if OneDrive is already signed in - great for users who are already syncing SharePoint libraries but not backing up their Data. 

Ideally, setup a monitor for the text "OneDrive" in the users folder path so you can trigger this off of a failed state in a monitor - use it as a self healing action.